Allocate and deallocate ips one-by-one

+++ bin/reserve_ips.py
@@ -16,6 +16,23 @@ defined in the shell
   - OS_TENANT_NAME: name of the tenant ips are going to be allocated
   - OS_PUBLIC_NET_ID: id of the net to reserve the ips from
 
+Under some circumstances, because of the way the neutron allocate and
+deallocated ips, it may be needed to run the script two times.
+The script allocate and release all the ips in the pool one-by-one
+taking track of them, so when an already allocated ip is extracted
+from the pool, the script knows the all the ips have been seen and
+that it can stop. Let say that the ip the script starts allocating
+from is 172.15.77.32 and that for some reason ip 172.15.74.0(that
+should be reserved) has been allocated and released, in that case
+that ip won't be recycled untill the pool will be exhausted and
+so won't be reserved by this script untill the next run.
+For this reason we should run this script two times and check
+that all the wanted ips have been reserved(The script's output included
+the reserved ips in the current run and all the ips allocated in the
+tenant).
+Note(rods): the script doesn't remove reserved ips already allocated
+so it's safe to run it more then once.
+
 """
 
 import os
@@ -49,34 +66,39 @@ def main():
 
     body = {'floatingip': {'floating_network_id': public_network_id}}
 
-    allocated = []
+    allocated = {}
+    reserved = []
 
-    print '\nAllocating temporary floating ips ',
+    special_ips = ['0', '1', '255']
+    keep_going = True
 
+    print
+    print 'Start ',
     try:
-        while True:
+        while keep_going:
+
             f_ip = neutron_client.create_floatingip(body).get('floatingip')
-            allocated.append(f_ip)
-            print_dot(len(allocated))
-    except neutron_exceptions.IpAddressGenerationFailureClient:
-        print 'Done'
-        print '\nFloaters allocated: %d' % len(allocated)
-    except Exception as err:
-        print
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        traceback.print_exception(exc_type, exc_value, exc_tb)
+            # we allocate and release all the ips in the pool one-by-one,
+            # so we can't count on the pool to be exhausted to know when
+            # the work is done. Let's keep a dict of seen ips and stop
+            # at the first ip seen more than once
+            # Note(rods):
+            #   We use a dict(with the ips as keys because) instead of just
+            #   a list of ips because the search for and element in the dict
+            #   should be much faster
 
-        print '\nFloaters allocated: %d' % len(allocated)
-        print '\nCleaning up all the allocated floaters ',
+            if f_ip['floating_ip_address'] in allocated:
+                keep_going = False
 
-        deallocated = 0
+            # keep track of the seen ips
+            allocated[f_ip['floating_ip_address']] = f_ip
 
-        for f_ip in allocated:
+            if f_ip['floating_ip_address'].split('.')[-1] in special_ips:
+                reserved.append(f_ip)
+                continue
             for _ in range(3):
                 try:
                     neutron_client.delete_floatingip(f_ip['id'])
-                    deallocated += 1
-                    print_dot(deallocated)
                     break
                 except Exception as err:
                     continue
@@ -84,35 +106,39 @@ def main():
                 print ('Floating ip %s has not been removed' %
                        f_ip['floating_ip_address'])
 
+            print_dot(len(allocated))
+
         print 'Done'
 
-        print '\nFloaters deallocated: %d' % deallocated
+    except Exception as err:
+
+        print
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        traceback.print_exception(exc_type, exc_value, exc_tb)
+
+        print '\nCleaning up all the allocated floaters ',
+
+        for f_ip in allocated.values():
+            # don't remove reserved ips already allocated
+            if f_ip['floating_ip_address'].split('.')[-1] not in special_ips:
+                for _ in range(3):
+                    try:
+                        neutron_client.delete_floatingip(f_ip['id'])
+                        print_dot(len(allocated))
+                        break
+                    except neutron_exceptions.NotFound as err:
+                        break
+                    except Exception as err:
+                        continue
+                else:
+                    print ('Floating ip %s has not been removed' %
+                           f_ip['floating_ip_address'])
+
+        print 'Done'
+
         return 1
 
-
-    print '\nDeallocate temporary floating ips ',
-
-    deallocated = 0
-    reserved = []
-
-    for f_ip in allocated:
-        if f_ip['floating_ip_address'].split('.')[-1] in ['0', '1', '255']:
-            reserved.append(f_ip)
-            continue
-        for _ in range(3):
-            try:
-                neutron_client.delete_floatingip(f_ip['id'])
-                deallocated += 1
-                print_dot(deallocated)
-                break
-            except Exception as err:
-                continue
-            else:
-                print ('Floating ip %s has not been removed' %
-                       f_ip['floating_ip_address'])
-    print 'Done'
-
-    print '\nFloaters deallocated: %d' % deallocated
+    print '\nFloaters allocated/deallocated %d' % len(allocated)
     print 'Floaters reserved: %d' % len(reserved)
 
     if reserved:
@@ -121,7 +147,10 @@ def main():
         for f_ip in reserved:
             print f_ip['floating_ip_address']
 
-    print ''
+    print '\nIps in the tenant:'
+    for f_ip in neutron_client.list_floatingips().get('floatingips', []):
+        print f_ip['floating_ip_address']
+
 
 if __name__ == '__main__':
     sys.exit(main())
